+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## [4.0.3](https://github.com/jayree/sfdx-jayree-plugin/compare/v4.0.2...v4.0.3) (2021-08-21)
+
+
+### Bug Fixes
+
+* update dependencies ([6f4b18b](https://github.com/jayree/sfdx-jayree-plugin/commit/6f4b18be647dbaabe7c38ce46d28a67718e2937d))
+* **gitdiff:** rename skipped to unchanged ([5a32c97](https://github.com/jayree/sfdx-jayree-plugin/commit/5a32c97f4ddbea1f16a8f1f215cffe369b53633e))
+* **gitdiff:** use lazy renderer option to fix sync issues ([495e9fe](https://github.com/jayree/sfdx-jayree-plugin/commit/495e9fedc65d340038e805820c6e082701d5861e))
+
+## [4.0.2](https://github.com/jayree/sfdx-jayree-plugin/compare/v4.0.1...v4.0.2) (2021-08-08)
+
+
+### Bug Fixes
+
+* **git:diff:** update registry, fix skip tasks ([bb74d76](https://github.com/jayree/sfdx-jayree-plugin/commit/bb74d7633c54897f6c164f6888dfe5f17befdde4))
+
+## [4.0.1](https://github.com/jayree/sfdx-jayree-plugin/compare/v4.0.0...v4.0.1) (2021-08-03)
+
+
+### Bug Fixes
+
+* **git:diff:** fix if process.emitWarning is called ([d0332ba](https://github.com/jayree/sfdx-jayree-plugin/commit/d0332bae087fc49f583be9b0a82ac4f343dba731))
+* **manifest:cleanup:** creates cleanup manifest template if file doesn't exist ([b6ae5e0](https://github.com/jayree/sfdx-jayree-plugin/commit/b6ae5e0e152a8fd4112b23274451bf003694b15d))
+* **manifest:generate:** fix if targetpath doesn't exist ([49a9937](https://github.com/jayree/sfdx-jayree-plugin/commit/49a9937d7f8ba1376ef9129bdb5a6ee4c2254e8b))
+
+# [4.0.0](https://github.com/jayree/sfdx-jayree-plugin/compare/v3.12.1...v4.0.0) (2021-08-01)
+
+
+### Bug Fixes
+
+* **hooks:** fix if result is empty after success filter ([5193185](https://github.com/jayree/sfdx-jayree-plugin/commit/51931851495cf34be61fce8092f8fd6cf68be39f))
+
+
+### Features
+
+* faster xml processing ([58bda42](https://github.com/jayree/sfdx-jayree-plugin/commit/58bda422aef53cb539ad6680e3b8d94e88037cf7))
+* hook support for plugin-source plugin ([fcb840b](https://github.com/jayree/sfdx-jayree-plugin/commit/fcb840b79d49d5739a87781425b98de44e5ce1db))
+* new lightning fast version of git:diff based on the  SDR library ([859e676](https://github.com/jayree/sfdx-jayree-plugin/commit/859e6763153ee2676117375b93bb17d474d29bd5))
+* remove deprecated commands ([009fd30](https://github.com/jayree/sfdx-jayree-plugin/commit/009fd30f132c6c90b1bd33a36fb237fee2108034))
+
+
+### BREAKING CHANGES
+
+* remove source:all and :full, as the functionality is replaced by the retrieve hooks
+
 ## [3.12.1](https://github.com/jayree/sfdx-jayree-plugin/compare/v3.12.0...v3.12.1) (2021-07-20)